Create task-3
For this project, I worked with the Iris flower dataset to build a classification model that predicts the species of a flower based on its sepal and petal measurements. I started to understand the given  dataset, then applied preprocessing and visualizations to identify patterns. This project helped me practice data analysis, visualization.